frontend-admin: display and basic interact with report

+++ backend-api/src/entities/reports.entity.ts
@@ -39,7 +39,7 @@ export enum ReportStatus {
 
 @Entity('reports')
 export class Report {
-  @PrimaryGeneratedColumn('increment', { type: 'bigint' })
+  @PrimaryGeneratedColumn('increment', { type: 'int' })
   id: number;
 
   @Column({
